requirements.txt
asgiref==3.7.2
Django==4.2.2
django-autoslug==1.9.9
django-debug-toolbar==4.1.0
django-widget-tweaks==1.4.12
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
